Adding Navbar.js
